+++ engine/src/eval/material_diff.rs
@@ -17,10 +17,10 @@ fn black_piece_value(game: &Game) -> Eval {
 
 fn count_piece_values(pieces: &PlayerPieces) -> Eval {
     Eval(
-        i32::from(pieces.pawns.count()) * PieceKind::value_of(PieceKind::Pawn)
-            + i32::from(pieces.knights.count()) * PieceKind::value_of(PieceKind::Knight)
-            + i32::from(pieces.bishops.count()) * PieceKind::value_of(PieceKind::Bishop)
-            + i32::from(pieces.rooks.count()) * PieceKind::value_of(PieceKind::Rook)
-            + i32::from(pieces.queens.count()) * PieceKind::value_of(PieceKind::Queen),
+        i16::from(pieces.pawns.count()) * PieceKind::value_of(PieceKind::Pawn)
+            + i16::from(pieces.knights.count()) * PieceKind::value_of(PieceKind::Knight)
+            + i16::from(pieces.bishops.count()) * PieceKind::value_of(PieceKind::Bishop)
+            + i16::from(pieces.rooks.count()) * PieceKind::value_of(PieceKind::Rook)
+            + i16::from(pieces.queens.count()) * PieceKind::value_of(PieceKind::Queen),
     )
 }
